Moved Planner section components to subfolder
