+++ src/components/sections/InventoryControls.vue
