+++ src/components/sections/SelectedOperators.vue
